AB#1376 Extract Yup schema for coordinates and apply it to custom fields of type geo_location
